requirements.txt
# latest versions

setuptools>=65.0.0
wheel
streamlit==1.45.1
pandas==1.5.3
numpy==1.24.3
pillow==9.4.0
